Refactored utility functions.
Moved commonly used utility functions to dedicated files
